+++ lib/src/customization/calendar_style.dart
@@ -29,6 +29,9 @@ class CalendarStyle {
   /// Style of foreground Text for holidays outside of current month.
   final TextStyle outsideHolidayStyle;
 
+  /// Style of foreground Text for days outside of `startDay` - `endDay` Date range.
+  final TextStyle unavailableStyle;
+
   /// Background Color of selected day.
   final Color selectedColor;
 
@@ -69,6 +72,11 @@ class CalendarStyle {
   /// * `false` - Today will have higher priority than SelectedDay
   final bool renderSelectedFirst;
 
+  /// Specifies if event markers rendered for a day cell can overflow cell's boundaries.
+  /// * `true` - Event markers will be drawn over the cell boundaries
+  /// * `false` - Event markers will not be drawn over the cell boundaries and will be clipped if they are too big
+  final bool canEventMarkersOverflow;
+
   const CalendarStyle({
     this.weekdayStyle = const TextStyle(),
     this.weekendStyle = const TextStyle(color: const Color(0xFFF44336)), // Material red[500]
@@ -78,6 +86,7 @@ class CalendarStyle {
     this.outsideStyle = const TextStyle(color: const Color(0xFF9E9E9E)), // Material grey[500]
     this.outsideWeekendStyle = const TextStyle(color: const Color(0xFFEF9A9A)), // Material red[200]
     this.outsideHolidayStyle = const TextStyle(color: const Color(0xFFEF9A9A)), // Material red[200]
+    this.unavailableStyle = const TextStyle(color: const Color(0xFFBFBFBF)),
     this.selectedColor = const Color(0xFF5C6BC0), // Material indigo[400]
     this.todayColor = const Color(0xFF9FA8DA), // Material indigo[200]
     this.markersColor = const Color(0xFF263238), // Material blueGrey[900]
@@ -89,5 +98,6 @@ class CalendarStyle {
     this.markersMaxAmount = 4,
     this.outsideDaysVisible = true,
     this.renderSelectedFirst = true,
+    this.canEventMarkersOverflow = false,
   });
 }